changing the name of canopy metrics as it is now very divergent and a pain to rebase
